Export de fonctions de sélection des actions dans un nouveau hook
